requirements.txt
fastapi==0.88.0
uvicorn==0.21.1
psutil==5.9.4
pycryptodome==3.17
pydantic==1.10.7
